chore: move hooks to page
