+++ packages/apps/human-app/frontend/src/modules/worker/email-verification/hooks/index.ts
@@ -1,4 +1,4 @@
-export * from './use-email-verification';
+export * from './use-email-verification-query';
 export * from './use-email-verification-token';
 export * from './use-resend-email';
 export * from './use-resend-email-router-params';
