--- packages/apps/human-app/frontend/src/modules/worker/email-verification/hooks/use-email-verification.ts
@@ -1,18 +1,0 @@
-import { useVerifyEmailQuery } from '@/modules/worker/services/email-verification';
-import { getErrorMessageForError } from '@/shared/errors';
-
-export function useEmailVerification(token: string) {
-  const {
-    error: emailVerificationError,
-    isError: isEmailVerificationError,
-    isPending: isEmailVerificationPending,
-  } = useVerifyEmailQuery({ token });
-
-  return {
-    errorMsg: emailVerificationError
-      ? getErrorMessageForError(emailVerificationError)
-      : null,
-    isEmailVerificationError,
-    isEmailVerificationPending,
-  };
-}
